// frontend/src/types/types.ts
export interface INewsFeed {
  items: IPost[];
  hasNewFeed: boolean;
  offset: number;
}

export interface IPost {
  id: string;
  privacy: "public" | "private" | "follower";
  photos: Record<string, any>[];
  comments: any[];
  description: string;
  likes: any[];
  author: IUser;
  commentsCount: number;
  likesCount: number;
  isBookmarked: boolean;
  isLiked: boolean;
  isOwnPost: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface IUser {
  id: string;
  username: string;
  fullName: string;
  email: string;
  profilePicture?: string | Record<string, any>;
  isEmailValidated: boolean;
  [prop: string]: any;
}